// my-app/src/types/IFeedItem.tsx
export interface IFeedItem {
  id: number;
  title: string;
  points: number | null;
  user: string | null;
  time: number;
  time_ago: string;
  content: string;
  deleted?: boolean;
  dead?: boolean;
  type: string;
  url?: string;
  domain?: string;
  comments: IFeedItem[]; // Comments are items too
  level: number;
  comments_count: number;
}